Move tests utils elsewhere
